Refactor charts tests
Create base class for charts with:
- shared imports
- default values setup
- tearDown function
