+++ Charts/ChartsTests/LineChartTests.swift
@@ -1,8 +1,6 @@
-import XCTest
-import FBSnapshotTestCase
 @testable import Charts
 
-class LineChartTests: FBSnapshotTestCase
+class LineChartTests: BaseChartTest
 {
     
     var chart: LineChartView!
@@ -11,17 +9,6 @@ class LineChartTests: FBSnapshotTestCase
     override func setUp()
     {
         super.setUp()
-        
-        // Set to `true` to re-capture all snapshots
-        self.recordMode = false
-        
-        // Sample data
-        let values: [Double] = [8, 104, 81, 93, 52, 44, 97, 101, 75, 28,
-            76, 25, 20, 13, 52, 44, 57, 23, 45, 91,
-            99, 14, 84, 48, 40, 71, 106, 41, 45, 61]
-        
-        var entries: [ChartDataEntry] = Array()
-        var xValues: [String] = Array()
         
         for (i, value) in values.enumerate()
         {
@@ -33,12 +20,6 @@ class LineChartTests: FBSnapshotTestCase
         
         chart = LineChartView(frame: CGRectMake(0, 0, 480, 350))
         chart.data = LineChartData(xVals: xValues, dataSet: dataSet)
-    }
-    
-    override func tearDown()
-    {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
     }
     
     func testDefaultValues()
